separate out bidder and notifier
