+++ tools/bidder-bot/notifier/notifier.go
@@ -2,19 +2,11 @@ package notifier
 
 import (
 	"context"
-	"encoding/hex"
 	"errors"
-	"io"
 	"log/slog"
-	"math/big"
 	"time"
 
-	"github.com/ethereum/go-ethereum/common"
-	"github.com/ethereum/go-ethereum/core/types"
-	bidderapiv1 "github.com/primev/mev-commit/p2p/gen/go/bidderapi/v1"
-	debugapiv1 "github.com/primev/mev-commit/p2p/gen/go/debugapi/v1"
 	notificationsapiv1 "github.com/primev/mev-commit/p2p/gen/go/notificationsapi/v1"
-	"github.com/primev/mev-commit/x/keysigner"
 )
 
 const (
@@ -26,48 +18,20 @@ var (
 	ErrUnexpectedTopic = errors.New("unexpected msg topic")
 )
 
-var nowFunc = time.Now
-
-type L1Client interface {
-	PendingNonceAt(ctx context.Context, account common.Address) (uint64, error)
-	BlockByNumber(ctx context.Context, number *big.Int) (*types.Block, error)
-	BlockNumber(ctx context.Context) (uint64, error)
-	HeaderByNumber(ctx context.Context, number *big.Int) (*types.Header, error)
-	ChainID(ctx context.Context) (*big.Int, error)
-}
-
 type Notifier struct {
 	logger              *slog.Logger
-	bidderClient        bidderapiv1.BidderClient
-	topologyClient      debugapiv1.DebugServiceClient
 	notificationsClient notificationsapiv1.NotificationsClient
-	l1Client            L1Client
-	signer              keysigner.KeySigner
-	gasTipCap           *big.Int
-	gasFeeCap           *big.Int
-	proposerChan        chan<- *upcomingProposer
+	proposerChan        chan<- *UpcomingProposer
 }
 
 func NewNotifier(
 	logger *slog.Logger,
-	bidderClient bidderapiv1.BidderClient,
-	topologyClient debugapiv1.DebugServiceClient,
 	notificationsClient notificationsapiv1.NotificationsClient,
-	l1Client L1Client,
-	signer keysigner.KeySigner,
-	gasTipCap *big.Int,
-	gasFeeCap *big.Int,
-	proposerChan chan<- *upcomingProposer,
+	proposerChan chan<- *UpcomingProposer,
 ) *Notifier {
 	return &Notifier{
 		logger:              logger,
-		bidderClient:        bidderClient,
-		topologyClient:      topologyClient,
 		notificationsClient: notificationsClient,
-		l1Client:            l1Client,
-		signer:              signer,
-		gasTipCap:           gasTipCap,
-		gasFeeCap:           gasFeeCap,
 		proposerChan:        proposerChan,
 	}
 }
@@ -76,8 +40,7 @@ func (b *Notifier) Start(ctx context.Context) <-chan struct{} {
 	done := make(chan struct{})
 	go func() {
 		defer close(done)
-
-		lastMsg := nowFunc()
+		lastMsg := time.Now()
 	RESTART:
 		sub, err := b.notificationsClient.Subscribe(ctx, &notificationsapiv1.SubscribeRequest{
 			Topics: []string{upcomingProposerTopic},
@@ -88,9 +51,10 @@ func (b *Notifier) Start(ctx context.Context) <-chan struct{} {
 		}
 		b.logger.Debug("subscribed to notifications", "topics", []string{upcomingProposerTopic})
 
-		if time.Since(lastMsg) > 15*time.Minute {
-			b.logger.Error("no messages received for 15 minutes, closing subscription")
-			return
+		// TODO: Address how plausible this is on mainnet with current opt-in numbers
+		if time.Since(lastMsg) > 1*time.Hour {
+			b.logger.Warn("no messages received for 1 hour, restarting subscription")
+			goto RESTART
 		}
 
 		for {
@@ -106,9 +70,7 @@ func (b *Notifier) Start(ctx context.Context) <-chan struct{} {
 				b.logger.Error("failed to receive message", "error", err)
 				goto RESTART
 			}
-
-			lastMsg = nowFunc()
-
+			lastMsg = time.Now()
 			b.logger.Info("received message", "topic", msg.Topic)
 
 			upcomingProposer, err := parseUpcomingProposer(msg)
@@ -116,7 +78,6 @@ func (b *Notifier) Start(ctx context.Context) <-chan struct{} {
 				b.logger.Error("failed to parse upcoming proposer", "error", err)
 				continue
 			}
-
 			select {
 			case b.proposerChan <- upcomingProposer:
 				b.logger.Debug("sent upcoming proposer to bidder worker", "proposer", upcomingProposer)
@@ -128,97 +89,13 @@ func (b *Notifier) Start(ctx context.Context) <-chan struct{} {
 	return done
 }
 
-func (b *Notifier) Bid(
-	ctx context.Context,
-	bidAmount *big.Int,
-	rawTx string,
-) error {
-	topo, err := b.topologyClient.GetTopology(ctx, &debugapiv1.EmptyMessage{})
-	if err != nil {
-		b.logger.Error("failed to get topology", "error", err)
-		return err
-	}
-
-	providers := topo.Topology.Fields["connected_providers"].GetListValue()
-	if providers == nil || len(providers.Values) == 0 {
-		return errors.New("no connected providers")
-	}
-
-	tx, err := b.SelfETHTransfer()
-	if err != nil {
-		b.logger.Error("failed to create self ETH transfer transaction", "error", err)
-		return err
-	}
-
-	// TODO: sanity check tx serialization
-	txBytes, err := tx.MarshalBinary()
-	if err != nil {
-		b.logger.Error("failed to marshal transaction", "error", err)
-		return err
-	}
-	txString := hex.EncodeToString(txBytes)
-
-	blkNumber, err := b.l1Client.BlockNumber(ctx)
-	if err != nil {
-		b.logger.Error("failed to get block number", "error", err)
-		return err
-	}
-
-	pc, err := b.bidderClient.SendBid(ctx, &bidderapiv1.Bid{
-		TxHashes:            []string{},
-		Amount:              bidAmount.String(),
-		BlockNumber:         int64(blkNumber + 1),
-		DecayStartTimestamp: nowFunc().UnixMilli(),
-		DecayEndTimestamp:   nowFunc().Add(12 * time.Second).UnixMilli(),
-		RevertingTxHashes:   []string{},
-		RawTransactions:     []string{txString},
-		SlashAmount:         big.NewInt(0).String(), // TODO: determine slash amount
-	})
-	if err != nil {
-		b.logger.Error("failed to send bid", "error", err)
-		return err
-	}
-
-	commitments := make([]*bidderapiv1.Commitment, 0)
-
-	ctx, cancel := context.WithTimeout(ctx, 12*time.Second)
-	defer cancel()
-
-	for {
-		select {
-		case <-ctx.Done():
-			return ctx.Err()
-		default:
-		}
-
-		msg, err := pc.Recv()
-		if err != nil {
-			if errors.Is(err, io.EOF) {
-				break
-			}
-			b.logger.Error("failed to receive commitment", "error", err)
-			return err
-		}
-
-		commitments = append(commitments, msg)
-
-		if len(commitments) == len(providers.Values) {
-			b.logger.Info("all commitments received")
-			return nil
-		} else {
-			b.logger.Warn(
-				"not all commitments received",
-				"received", len(commitments),
-				"expected", len(providers.Values),
-			)
-		}
-	}
-	b.logger.Info("bid succeeded, but not all commitments received", "commitments", len(commitments))
-
-	return nil
+type UpcomingProposer struct {
+	BLSKey string
+	Epoch  uint64
+	Slot   uint64
 }
 
-func parseUpcomingProposer(msg *notificationsapiv1.Notification) (*upcomingProposer, error) {
+func parseUpcomingProposer(msg *notificationsapiv1.Notification) (*UpcomingProposer, error) {
 	if msg.Topic != upcomingProposerTopic {
 		return nil, ErrUnexpectedTopic
 	}
@@ -247,15 +124,9 @@ func parseUpcomingProposer(msg *notificationsapiv1.Notification) (*upcomingPropo
 		return nil, errors.New("failed to parse slot")
 	}
 
-	return &upcomingProposer{
+	return &UpcomingProposer{
 		BLSKey: blsKey,
 		Epoch:  uint64(epochVal),
 		Slot:   uint64(slotVal),
 	}, nil
-}
-
-type upcomingProposer struct {
-	BLSKey string
-	Epoch  uint64
-	Slot   uint64
 }
